chore: calculation 패키지 분리
