+++ src/main/java/com/example/yeobee/core/expense/domain/CustomExpenseRepository.java
@@ -1,8 +1,8 @@
 package com.example.yeobee.core.expense.domain;
 
+import com.example.yeobee.core.calculation.domain.CalculationResult;
 import com.example.yeobee.core.expense.dto.request.ExpenseListFilter;
 import com.example.yeobee.core.expense.dto.response.ExpenseListRetrieveResponseDto;
-import com.example.yeobee.core.trip.domain.CalculationResult;
 import java.util.List;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
